requirements.txt
langchain
langgraph
python-dotenv
langchain[google-genai]
fastapi
uvicorn[standard]
pymongo
qdrant-client
pymupdf